+++ HISTORY.md
@@ -122,3 +122,4 @@
 
 ## 0.4.9
 - update `getProvider` function to pass RPL urls
+- copy text in `<AddressLink />` component